Split up plugin types
